Pull relevant functions and documentations out of Tate models and into Literature models
