+++ experimental/FTheoryTools/docs/doc.main
@@ -4,5 +4,6 @@
       "weierstrass.md",
       "tate.md",
       "hypersurface.md",
+      "literature.md",
    ],
 ]
