+++ experimental/FTheoryTools/src/LiteratureModels/auxiliary.jl
