+++ experimental/FTheoryTools/src/TateModels/attributes.jl
@@ -15,7 +15,7 @@
 Return the Tate section ``a_1``.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> tate_section_a1(t)
@@ -31,7 +31,7 @@ tate_section_a1(t::GlobalTateModel) = t.tate_a1
 Return the Tate section ``a_2``.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> tate_section_a2(t)
@@ -47,7 +47,7 @@ tate_section_a2(t::GlobalTateModel) = t.tate_a2
 Return the Tate section ``a_3``.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> tate_section_a3(t)
@@ -63,7 +63,7 @@ tate_section_a3(t::GlobalTateModel) = t.tate_a3
 Return the Tate section ``a_4``.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> tate_section_a4(t)
@@ -79,7 +79,7 @@ tate_section_a4(t::GlobalTateModel) = t.tate_a4
 Return the Tate section ``a_6``.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> tate_section_a6(t)
@@ -99,7 +99,7 @@ tate_section_a6(t::GlobalTateModel) = t.tate_a6
 Return the Tate polynomial of the global Tate model.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> tate_polynomial(t)
@@ -119,7 +119,7 @@ tate_polynomial(t::GlobalTateModel) = t.tate_polynomial
 Return the base space of the global Tate model.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> base_space(t)
@@ -138,7 +138,7 @@ end
 Return the ambient space of the global Tate model.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> ambient_space(t)
@@ -188,7 +188,7 @@ Return the Calabi-Yau hypersurface in the toric ambient space
 which defines the global Tate model.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> calabi_yau_hypersurface(t)
@@ -212,7 +212,7 @@ end
 Return the Weierstrass model which is equivalent to the given Tate model.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> weierstrass_model(t)
@@ -246,7 +246,7 @@ end
 Return the discriminant of the global Tate model.
 
 ```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
+julia> t = literature_model(arxiv_id = "1109.3454", equation = "3.1")
 Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
 
 julia> discriminant(t);
@@ -320,153 +320,4 @@ julia> singular_loci(t)[2]
 @attr Vector{<:Tuple{<:MPolyIdeal{<:MPolyRingElem}, Tuple{Int64, Int64, Int64}, String}} function singular_loci(t::GlobalTateModel)
   @req typeof(base_space(t)) <: ToricCoveredScheme "Singular loci of global Tate model currently only supported for toric varieties/schemes as base space"
   return singular_loci(weierstrass_model(t))
-end
-
-
-#####################################################
-# 7: Resolutions for a global Tate model
-#####################################################
-
-@doc raw"""
-    resolutions(t::GlobalTateModel)
-
-Return the list of all known resolutions for the global
-Tate model.
-
-```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
-Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
-
-julia> length(resolutions(t))
-1
-```
-"""
-function resolutions(t::GlobalTateModel)
-  @req has_attribute(t, :resolutions) "No resolutions known for this model"
-  return get_attribute(t, :resolutions)
-end
-
-
-#####################################################
-# 8: Attributes for literature Tate models
-#####################################################
-
-@doc raw"""
-    doi(t::GlobalTateModel)
-
-Return the `doi` of the publication which introduced
-the global Tate model in question. If no `doi` is
-known, an error is raised.
-
-```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
-Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
-
-julia> doi(t)
-"10.1016/j.nuclphysb.2011.12.013"
-```
-"""
-function doi(t::GlobalTateModel)
-  @req has_attribute(t, :doi) "No doi known for this model"
-  return get_attribute(t, :doi)
-end
-
-@doc raw"""
-    arxiv_id(t::GlobalTateModel)
-
-Return the `arxiv_id` of the publication which introduced
-the global Tate model in question. If no `arxiv_id` is
-known, an error is raised.
-
-```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
-Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
-
-julia> arxiv_id(t)
-"1109.3454"
-```
-"""
-function arxiv_id(t::GlobalTateModel)
-  @req has_attribute(t, :arxiv_id) "No arxiv identifier known for this model"
-  return get_attribute(t, :arxiv_id)
-end
-
-@doc raw"""
-    version(t::GlobalTateModel)
-
-Return the `version` of the arxiv publication which
-introduced the global Tate model in question. If no
-`version` is known, an error is raised.
-
-```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
-Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
-
-julia> version(t)
-"2"
-```
-"""
-function version(t::GlobalTateModel)
-  @req has_attribute(t, :version) "No version known for this model"
-  return get_attribute(t, :version)
-end
-
-@doc raw"""
-    equation_number(t::GlobalTateModel)
-
-Return the `equation_number` in which the model was introduced
-in the publication on our record (either specified by
-`doi` or, if uploaded to the arxiv, the `arxiv_id` and
-`version`). If no `equation_number` is known, an error is raised.
-
-```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
-Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
-
-julia> equation_number(t)
-"3.1"
-```
-"""
-function equation_number(t::GlobalTateModel)
-  @req has_attribute(t, :equation_number) "No equation number known for this model"
-  return get_attribute(t, :equation_number)
-end
-
-@doc raw"""
-    description(t::GlobalTateModel)
-
-Return the `description` of the model in question.
-If no `description` is known, an error is raised.
-
-```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
-Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
-
-julia> description(t)
-"SU(5)xU(1) restricted Tate model"
-```
-"""
-function description(t::GlobalTateModel)
-  @req has_attribute(t, :description) "No description known for this model"
-  return get_attribute(t, :description)
-end
-
-@doc raw"""
-    link(t::GlobalTateModel)
-
-Return the `link` (formatted as string) to the online
-version of the paper which introduced the model in question.
-If no link is known, an error is raised.
-
-```jldoctest
-julia> t = literature_tate_model(arxiv_id = "1109.3454", equation = "3.1")
-Global Tate model over a not fully specified base -- SU(5)xU(1) restricted Tate model based on arXiv paper 1109.3454 Eq. (3.1)
-
-julia> link(t)
-"https://arxiv.org/abs/1109.3454v2"
-```
-"""
-function link(t::GlobalTateModel)
-  @req has_attribute(t, :link) "No link known for this model"
-  return get_attribute(t, :link)
 end
