+++ experimental/FTheoryTools/src/exports.jl
@@ -39,7 +39,7 @@ export hypersurface_model
 export set_hypersurface_equation
 export singular_loci
 export link
-export literature_tate_model
+export literature_model
 export resolutions
 export resolve
 export set_description
